requirements.txt
Pillow
PyInstaller
appdirs
